refactor module name and some typo
