+++ document-parsers/langchain4j-community-document-parser-llamaparse/src/main/java/dev/langchain4j/community/data/document/parser/llamaparse/LlamaParseResponse.java
@@ -1,4 +1,4 @@
-package dev.langchain4j.data.document.parser.llamaparse;
+package dev.langchain4j.community.data.document.parser.llamaparse;
 
 import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
 
@@ -7,7 +7,8 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 
 @JsonInclude(NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class LlamaParseMarkdownResponse {
+public class LlamaParseResponse {
 
-    public String markdown;
+    public String id;
+    public String status;
 }
