+++ document-parsers/langchain4j-community-document-parser-llamaparse/src/test/resources/files/sample.pdf
