+++ pom.xml
@@ -71,7 +71,7 @@
         <module>web-search-engines/langchain4j-community-web-search-engine-searxng</module>
 
         <!-- Integration of document parsers -->
-        <module>document-parsers/langchain4j-document-parser-llamaparse</module>
+        <module>document-parsers/langchain4j-community-document-parser-llamaparse</module>
 
         <!-- Spring Boot Starters -->
         <module>spring-boot-starters</module>
